@@ -3,11 +3,13 @@ import 'package:provider/provider.dart';
 import 'package:smarthouse/blocs/home_page_bloc.dart';
 import 'package:smarthouse/presentations/components/room_expand_tile.dart';
 import 'package:smarthouse/models/room.dart';
+import 'package:smarthouse/presentations/home_page/heading_tile.dart';
 import 'package:smarthouse/presentations/home_page/room_tile.dart';
 import 'package:smarthouse/providers/device_provider.dart';
 import 'package:smarthouse/providers/room_provider.dart';
 
 import '../../blocs/home_page_bloc.dart';
+import 'device_tile.dart';
 
 //class HomePage extends StatefulWidget {
 //  static const String routeName = "/";
@@ -145,27 +147,26 @@ class _HomePageState extends State<HomePage> {
   @override
   void initState() {
     super.initState();
-    Provider.of<RoomProvider>(context, listen: false).fetch();
     Provider.of<DeviceProvider>(context, listen: false).fetch();
+    Provider.of<RoomProvider>(context, listen: false).fetch();
   }
   @override
   Widget build(BuildContext context) {
     RoomProvider roomProvider =  Provider.of<RoomProvider>(context);
     DeviceProvider deviceProvider =  Provider.of<DeviceProvider>(context);
-    print(roomProvider);
     return Scaffold(
       backgroundColor: Theme.of(context).backgroundColor,
       appBar: AppBar(
         title: Text(
           "Smart house",
-          style: TextStyle(
-            fontSize: 25,
-          ),
         ),
       ),
-      body: ListView.builder(itemCount: roomProvider.rooms.length + deviceProvider.devices.length,itemBuilder: (ctx, index) {
-        return index < deviceProvider.devices.length ? RoomTile(room: roomProvider.rooms[index],) : Text(deviceProvider.devices[index - deviceProvider.devices.length].name);
-      }),
+      body: ListView(children: <Widget>[
+        HeadingTile("Rooms"),
+        ...roomProvider.rooms.map((room) => RoomTile(room)).toList(),
+        HeadingTile("Unassigned Devices"),
+        ...deviceProvider.unassginedDevices.map((device) => DeviceTile(device.name)).toList()
+      ],)
     );
   }
 }
