@@ -1,23 +1,21 @@
 import 'package:flutter/material.dart';
 import 'package:smarthouse/models/room.dart';
+import 'package:smarthouse/presentations/home_page/device_tile.dart';
 
 import '../room_page/room_page.dart';
 
-class RoomTile extends StatefulWidget {
+class RoomTile extends StatelessWidget {
   final Room room;
-  RoomTile({this.room});
-  @override
-  _RoomTileState createState() => _RoomTileState();
-}
-
-class _RoomTileState extends State<RoomTile> {
+  const RoomTile(this.room);
   @override
   Widget build(BuildContext context) {
-    return Card(
-      child: ListTile(
-        title: Text(widget.room.name),
-        onTap: () => Navigator.of(context).pushNamed(RoomPage.routeName),
-      ),
-    );
+    return room.deviceList.isEmpty ? ListTile(
+      title: Text(room.name),
+      onTap: () => Navigator.of(context).pushNamed(RoomPage.routeName),
+    ) : ExpansionTile(title: Text(room.name), children:
+      room.deviceList.map((device) => DeviceTile(device.name)).toList());
   }
 }
+
+
+  
